refactoring for merge cells.
